% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixpanel_io.R
\name{retrieve}
\alias{retrieve}
\alias{read}
\title{Retrieve and read MixPanel CSV files}
\usage{
retrieve(destination = R_user_dir("mixpanel", "cache"))

read(tbl)
}
\arguments{
\item{destination}{character(1) location of the local cache. The
default is the standard operating system location for R
packages.}

\item{tbl}{A tibble returned by \code{retrieve()}, typically filtered to
select a title of interest.}
}
\value{
\code{retrieve()} returns a tibble with columns year, month,
title, and path to the localized file. The year, month, and
title are parsed from the file name.

\code{read()} returns a single tibble.
}
\description{
Retrieve and read MixPanel CSV files
}
\details{
\code{retrieve()} visits the Google bucket where AnVIL MixPanel
data are stored, and retrieves all \verb{*.csv} files to a local
cache. Files already present in the cache are not
re-localized. You will be prompted to authenticate with an
account for which shared access to the \verb{MixPanel 2021} Google
Drive folder.

\code{read()} reads several csv files into tibbles, and then
binds rows into a single tibble. Column headers with a hyphen \code{-}, typically separating date ranges spanning the month, are
replaced with \code{DateRangeCount}.
}
